Add support for Flux
Add a flux "job shell plugin" which allows a Spindle session to
be started alongside the job. The plugin runs the Spindle frontend
on the shell rank 0, and starts the backend on all ranks.

The plugin should be installed in the default shell plugin path and
is activated when the `spindle` shell options is set, e.g. via

 flux mini run -o spindle OPTIONS... COMMAND...