Fixes hang introduced by logging in sp_init().
Logging initialization relies on environment variables, and
some of those variables originate in the flux shell.  Initiali-
zation prior to those values being transferred can result in a
hang.

This PR moves the transfer code into flux_plugin_init(), which is
the first spindle-provided callback to be executed.

Passes all runTests on Tuolumne.