Adds cachepaths and commpaths parameters.
Both are colon-separated lists of paths.  Not yet integrated into
flux parameter handling.

Splitting "location" in two and allowing for lists of paths
allows the file cache to be placed in a "remote" directory
(e.g., "rabbits") while accounting for runtime failure of those
directories (if the "rabbit" directory isn't mounted on a
particular node, Spindle falls back to the next direcotry in
the list).

Fifos and pipes are probably best kept in a more-local filesystem,
thus necessitating the split into two parameters.